requirements.txt
Django==4.1.4
djangorestframework==3.14.0
Jinja2==3.1.2
psycopg2==2.9.5
twilio==7.16.0
